### 2023-05-31 #5507 @editorialbot
[PRE REVIEW]: NEMSEER: A Python package for downloading and handling historical National Electricity Market forecast data produced by the Australian Energy Market Operator
**Submitting author:** @prakaa (<a href="http://orcid.org/0000-0002-2945-4757">Abhijith Prakash</a>)
**Repository:** https://github.com/UNSW-CEEM/NEMSEER
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.0.4
