### 2023-05-31 #5507 @prakaa
Hi @kyleniemeyer, thanks for your response and for taking the time to look at this.

I have tried to address some of these points below in the paper, but I hope this additional context below is useful:

- I think the [examples](https://nemseer.readthedocs.io/en/latest/examples.html) in the docs is where a lot of the value is for users.
  - They are focused on forecast errors as this is likely of interest to many users. Since the scope of the package is wider than power system and electricity market forecast errors, functions that explicitly calculate these have not been implemented in the package itself.
  - However the docs provided coded examples of how NEMSEER and sister packages can be used to calculate and characterise these errors
  - An example: [this online companion](https://github.com/oyurdakul/pesgm23) to a PESGM23 conference paper contains [scripts](https://github.com/oyurdakul/pesgm23/blob/main/stats/stats_24/self_comm_statistical_24_hr.py) that show that the researchers used code directly from the [energy price convergence example in the docs](https://nemseer.readthedocs.io/en/latest/examples/price_convergence_2021.html).
- Beyond the examples, the package's documentation also acts as a repository of otherwise disparate knowledge about the datasets.

Based on this, I think NEMSEER *makes addressing research challenges significantly better (faster, easier, simpler)* and assists researchers in *extracting knowledge from large data sets*. I think it's fair to say that a lot of this value is delivered not only through the package, but also through the accompanying documentation (with examples).
