### 2023-07-04 #5507 @prakaa
@kyleniemeyer thanks for letting me know, appreciate the JOSS editorial team is probably quite busy.

Potential reviewers:
James Foster (https://github.com/jd-foster) and Changlong Wang (https://github.com/changlongw)
