### 2023-07-05 d8b0531
update pre-commit hooks and link flake8 hook to github no gitlab
