+++ .pre-commit-config.yaml
@@ -3,7 +3,7 @@
 # See https://pre-commit.com/hooks.html for more hooks
 repos:
 -   repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.3.0
+    rev: v4.4.0
     hooks:
     -   id: trailing-whitespace # trims trailing whitespace
     -   id: end-of-file-fixer # files end with one newline
@@ -13,14 +13,14 @@ repos:
 # Runs isort, which sorts imports
 # Note that it is configured with black-compatible settings
 -   repo: https://github.com/pycqa/isort
-    rev: 5.10.1
+    rev: 5.12.0
     hooks:
     -   id: isort
         args: ["--profile", "black", "--filter-files"]
 
 # Runs black
 -   repo: https://github.com/psf/black
-    rev: 22.6.0
+    rev: 23.3.0
     hooks:
     -   id: black
         # It is recommended to specify the latest version of Python
@@ -30,8 +30,8 @@ repos:
         language_version: python3.10
 
 # Runs flake8
--   repo: https://gitlab.com/pycqa/flake8
-    rev: 3.9.2
+-   repo: https://github.com/pycqa/flake8
+    rev: 6.0.0
     hooks:
     -   id: flake8
 
