### 2023-07-05 bb4ac8f
add pre-commit hook installation to contributing
