### 2023-07-05 a0265f3
change license type and add licence for documentation
