### 2023-07-05 e92559e
update poetry deps, make nemseer compatible with 3.11 and add 3.11 to CI/CD
