### 2023-07-05 3104b5d
add CC BY logos to docs
