+++ .pre-commit-config.yaml
@@ -20,7 +20,7 @@ repos:
 
 # Runs black
 -   repo: https://github.com/psf/black
-    rev: 23.3.0
+    rev: 23.7.0
     hooks:
     -   id: black
         # It is recommended to specify the latest version of Python
