### 2023-07-25 e04b6c9
remove stringent dependencies and bump version
