### 2023-07-25 becf69e
ensure seconds can be parsed in datetime cols
