+++ src/nemseer/data_handlers.py
@@ -31,7 +31,7 @@ def _parse_datetime_cols(df: pd.DataFrame) -> pd.DataFrame:
     dt_cols = DATETIME_COLS
     dt_cols_present = dt_cols.intersection(set(df.columns.tolist()))
     for col in dt_cols_present:
-        df.loc[:, col] = pd.to_datetime(df[col], format=DATETIME_FORMAT)
+        df.loc[:, col] = pd.to_datetime(df[col], format=DATETIME_FORMAT + ":%S")
     return df
 
 
