### 2023-07-25 0099a4f
use mypy mirror and fix mypy suggestions in data_compilers
