### 2023-07-25 c1344c4
Add pre-commit ci badge
