+++ README.md
@@ -4,6 +4,7 @@
 [![Continuous Integration and Deployment](https://github.com/UNSW-CEEM/NEMSEER/actions/workflows/cicd.yml/badge.svg)](https://github.com/UNSW-CEEM/NEMSEER/actions/workflows/cicd.yml)
 [![Documentation Status](https://readthedocs.org/projects/nemseer/badge/?version=latest)](https://nemseer.readthedocs.io/en/latest/?badge=latest)
 [![codecov](https://codecov.io/gh/UNSW-CEEM/NEMSEER/branch/master/graph/badge.svg?token=BO69YSQIGI)](https://codecov.io/gh/UNSW-CEEM/NEMSEER)
+[![pre-commit.ci status](https://results.pre-commit.ci/badge/github/UNSW-CEEM/NEMSEER/master.svg)](https://results.pre-commit.ci/latest/github/UNSW-CEEM/NEMSEER/master)
 [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
 [![DOI](https://zenodo.org/badge/508154456.svg)](https://zenodo.org/badge/latestdoi/508154456)
 
