### 2023-07-28 672cedf
add polars to quick start in docs
