### 2023-07-28 137e37f
update nemseer examples with matplotlib styling
