### 2023-07-28 524490a
update nemseer examples with matplotlib styling
