+++ docs/source/examples/visualising_p5min_demand_forecasts.md
@@ -59,6 +59,12 @@ import plotly.io as pio
 import plotly.express as px
 ```
 
+## Plot styling
+
+```{code-cell} ipython3
+plt.style.use(Path("styling", "matplotlib_styling.mplstyle"))
+```
+
 ## Study times
 
 Here we'll define our datetime range that we're interested in:
