### 2023-07-31 faee21e
[pre-commit.ci] pre-commit autoupdate
updates:
- [github.com/pycqa/flake8: 6.0.0 → 6.1.0](https://github.com/pycqa/flake8/compare/6.0.0...6.1.0)
