+++ .pre-commit-config.yaml
@@ -31,7 +31,7 @@ repos:
 
 # Runs flake8
 -   repo: https://github.com/pycqa/flake8
-    rev: 6.0.0
+    rev: 6.1.0
     hooks:
     -   id: flake8
 
