+++ .pre-commit-config.yaml
@@ -38,7 +38,7 @@ repos:
 # Runs mypy
 # Uses local environment as opposed to mypy-mirror to enable access to dependencies
 -   repo: https://github.com/pre-commit/mirrors-mypy
-    rev: v1.5.0
+    rev: v1.5.1
     hooks:
     -   id: mypy
         #        name: mypy
