### 2023-09-26 #5507 @arfon
@mfleschutz @blthayer @amandadsmith – 👋 would any of you be willing to review this submission for JOSS? The submission under consideration is iNEMSEER: A Python package for downloading and handling historical National Electricity Market forecast data produced by the Australian Energy Market Operator: https://github.com/UNSW-CEEM/NEMSEER

As previous authors of somewhat similar packages, we think you might be able to provide a great review of this submission. Please let me know if you think you can help us out!

Many thanks
Arfon
