### 2023-10-12 #5883 @mfleschutz
So I am starting my proper review.

First thing is noticed when reading through the documentation:

In https://nemseer.readthedocs.io/en/latest/examples/visualising_p5min_demand_forecasts.html there are a lot of annoying warnings:

> WARNING: findfont: Generic family 'sans-serif' not found because none of the following families were found: Source Sans 3, Helvetica, Computer Modern Sans Serif, Avant Garde, sans-serif

@prakaa, can you get rid of them?
