### 2023-10-12 82f9098
remove plot styling from visualisation example to prevent warnings
