### 2023-10-12 #5883 @mfleschutz
- [ ] @prakaa, some of the References in the JOSS paper are no books but have no URL or DOI.
