### 2023-10-12 #5883 @mfleschutz
I use a MacBook Pro M1 to test. When running the `p5min_demand_forecast_error_2021.ipynb` the  `download_raw_data` works. However, calculating the forecast_error (in the cell containing the line `forecast_error = pd.concat(results, axis=0)`) throws the following error multiple times, so I stopped the calculation:

```raw
multiprocessing/process.py", line 108, in run
    self._target(*self._args, **self._kwargs)
  File ".../python3.9/multiprocessing/pool.py", line 114, in worker
    task = get()
  File ".../python3.9/multiprocessing/queues.py", line 368, in get
    return _ForkingPickler.loads(res)
AttributeError: Can't get attribute 'calculate_p5min_demand_forecast_error_simpler' on <module '__main__' (built-in)>
Traceback (most recent call last):
  File ".../python3.9/multiprocessing/process.py", line 315, in _bootstrap
    self.run()
  File ".../python3.9/multiprocessing/process.py", line 108, in run
    self._target(*self._args, **self._kwargs)
```
