### 2023-10-19 fa38792
add urls/dois to refs missing them
