+++ joss-paper/paper.bib
@@ -1,179 +1,184 @@
-@report{australianenergymarketcommissionReserveServicesNational2021,
-  title = {Reserve Services in the {{National Electricity Market}}, {{Directions Paper}}},
+@techreport{australianenergymarketcommissionReserveServicesNational2021,
+  title  = {Reserve services in the {National} {Electricity} {Market}, {Directions} {Paper}},
+  url    = {https://www.aemc.gov.au/sites/default/files/2021-01/Reserve%20services%20directions%20paper%20-%205.01.2021%20-%20FINAL.pdf},
   author = {{Australian Energy Market Commission}},
-  date = {2021},
+  year   = {2021}
 }
 
 @misc{australianenergymarketoperatorPreDispatch,
-  title = {Pre Dispatch},
+  title  = {Pre Dispatch},
+  url    = {https://aemo.com.au/energy-systems/electricity/national-electricity-market-nem/data-nem/market-management-system-mms-data/pre-dispatch},
   author = {{Australian Energy Market Operator}},
-  date = {2022},
-  urldate = {2022-05-24},
-  howpublished = {https://aemo.com.au/energy-systems/electricity/national-electricity-market-nem/data-nem/market-management-system-mms-data/pre-dispatch},
-  langid = {english}
+  date   = {2022}
 }
 
 @techreport{australianenergymarketoperatorPredispatchOperatingProcedure2021,
-  title = {Pre-Dispatch Operating Procedure},
-  author = {{Australian Energy Market Operator}},
-  year = {2021},
-  month = oct,
+  title   = {Pre-dispatch operating procedure},
+  url     = {https://www.aemo.com.au/-/media/files/electricity/nem/security_and_reliability/power_system_ops/procedures/so_op_3704-predispatch.pdf?la=en},
   urldate = {2022-05-24},
-  file = {/home/abi/Dropbox/zotero/Australian Energy Market Operator/australian_energy_market_operator_2021_pre-dispatch_operating_procedure.pdf}
+  author  = {{Australian Energy Market Operator}},
+  month   = oct,
+  year    = {2021}
 }
 
 @report{australianenergymarketoperatorPreDispatchSensitivities2021,
-  title = {Pre-{{Dispatch Sensitivities}}},
-  author = {{Australian Energy Market Operator}},
-  date = {2021-03},
-  url = {https://www.aemo.com.au/-/media/files/electricity/nem/security_and_reliability/dispatch/policy_and_process/pre-dispatch-sensitivities.pdf},
+  title   = {Pre-{{Dispatch Sensitivities}}},
+  url     = {https://www.aemo.com.au/-/media/files/electricity/nem/security_and_reliability/dispatch/policy_and_process/pre-dispatch-sensitivities.pdf},
   urldate = {2022-05-24},
+  author  = {{Australian Energy Market Operator}},
+  date    = {2021-03}
 }
 
 @techreport{australianenergymarketoperatorReliabilityStandardImplementation2020,
-  title = {Reliability {{Standard Implementation Guidelines}}},
-  author = {{Australian Energy Market Operator}},
-  year = {2020},
-  month = dec,
+  title   = {Reliability {Standard} {Implementation} {Guidelines}},
+  url     = {https://aemo.com.au/-/media/files/electricity/nem/planning_and_forecasting/rsig/reliability-standard-implementation-guidelines.pdf?la=en},
   urldate = {2022-08-10},
+  author  = {{Australian Energy Market Operator}},
+  month   = dec,
+  year    = {2020}
 }
 
 @article{cramtonElectricityMarketDesign2017,
-  title = {Electricity Market Design},
[…]
