+++ joss-paper/paper.bib
@@ -137,7 +137,7 @@
   urldate   = {2023-03-13},
   publisher = {The Apache Software Foundation},
   month     = mar,
-  year      = {2023},
+  year      = {2023}
 }
 
 @software{prakashNEMStorageUnderUncertainty2022,
@@ -148,7 +148,7 @@
   publisher = {GitHub},
   journal   = {GitHub repository},
   url       = {https://github.com/prakaa/NEMStorageUnderUncertainty},
-  urldate   = {2023-03-14},
+  urldate   = {2023-03-14}
 }
 
 @misc{reback2020pandas,
@@ -162,26 +162,27 @@
 
 
 @incollection{roquesEvolutionEuropeanModel2021,
-	title = {The evolution of the {European} model for electricity markets},
-	isbn = {978-1-78897-994-8},
-	url = {https://www.elgaronline.com/display/edcoll/9781788979948/9781788979948.00017.xml},
-	language = {en\_US},
-	urldate = {2023-03-13},
-	booktitle = {Handbook on {Electricity} {Markets}},
-	publisher = {Edward Elgar Publishing},
-	author = {Roques, Fabien},
-	month = nov,
-	year = {2021},
-	pages = {308--330},
+  title     = {The evolution of the {European} model for electricity markets},
+  isbn      = {978-1-78897-994-8},
+  url       = {https://www.elgaronline.com/display/edcoll/9781788979948/9781788979948.00017.xml},
+  doi       = {10.4337/9781788979955.00017},
+  language  = {en\_US},
+  urldate   = {2023-03-13},
+  booktitle = {Handbook on {Electricity} {Markets}},
+  publisher = {Edward Elgar Publishing},
+  author    = {Roques, Fabien},
+  month     = nov,
+  year      = {2021},
+  pages     = {308--330}
 }
 
 @software{yurdakulOnlineCompanionRiskAverse2023,
-  title    = {Online {{Companion}} to "{{Risk-Averse Self-Scheduling}} of {{Storage}} in {{Decentralized Markets}}"},
-  author   = {Yurdakul, Ogün and Billimoria, Farhad},
-  date     = {2023-02-23T14:16:45Z},
-  origdate = {2022-10-15T15:22:25Z},
-  url = {https://github.com/oyurdakul/pesgm23},
-  urldate = {2023-03-14},
+  title     = {Online {{Companion}} to "{{Risk-Averse Self-Scheduling}} of {{Storage}} in {{Decentralized Markets}}"},
+  author    = {Yurdakul, Ogün and Billimoria, Farhad},
+  date      = {2023-02-23T14:16:45Z},
+  origdate  = {2022-10-15T15:22:25Z},
+  url       = {https://github.com/oyurdakul/pesgm23},
+  urldate   = {2023-03-14},
   publisher = {GitHub},
-  journal = {GitHub repository},
+  journal   = {GitHub repository}
 }
