+++ joss-paper/paper.md
@@ -54,14 +54,14 @@ AEMO publicly releases data from five of its operational ahead-of-time processes
 - Short Term Projected Assessment of System Adequacy
 - Medium Term Projected Assessment of System Adequacy
 
-However, significant effort and prerequisite knowledge is required to obtain and process this data for analysis. Firstly, a user must be familiar with how AEMO's data repositories are organised. Secondly, a user must have knowledge of what type of data each ahead-of-time process generates (i.e. the range of tables and columns available), and of each process' lookahead horizon (i.e. for a given time at which the process is *run*, how many periods into the future are *forecasted*?). Finally, a user must download, unzip and clean CSV files  before being able to load and handle tables of interest using data analysis tools.
+However, significant effort and prerequisite knowledge is required to obtain and process this data for analysis. Firstly, a user must be familiar with how AEMO's data repositories are organised. Secondly, a user must have knowledge of what type of data each ahead-of-time process generates (i.e. the range of tables and columns available), and of each process' lookahead horizon (i.e. for a given time at which the process is *run*, how many periods into the future are *forecasted*?). Finally, a user must download, unzip and clean CSV files before being able to load and handle tables of interest using data analysis tools.
 
 `NEMSEER` solves these issues by:
 
 1. Providing learning resources and references (via the README and a [glossary](https://nemseer.readthedocs.io/en/latest/glossary.html) in the documentation) that unpack what each ahead-of-time process does and what data they offer.
 2. Making it easier to download and handle this data. `NEMSEER` can inform the user of the date range of available data, which data tables are available and even generate the appropriate range of [*run* times](https://nemseer.readthedocs.io/en/latest/glossary.html#term-run-times) for a set of [*forecasted* times](https://nemseer.readthedocs.io/en/latest/glossary.html#term-forecasted-times) that a user is interested in. Once a user queries a subset of data, `NEMSEER` will download, unzip and process the CSV files into `pandas` or `xarray` data structures.
 
-Furthermore, the package documentation contains examples (with Python code) that show how users can analyse demand forecast errors and energy price convergence using pre-dispatch demand and price forecast data (obtained using `NEMSEER`) and historical *actual* NEM system and market data (obtained using `NEMOSIS`) [@gormanNEMOSISNEMOpen2018]. \autoref{fig:p5demandforecasterror} is an output of one such [example](https://nemseer.readthedocs.io/en/latest/examples/p5min_demand_forecast_error_2021.html#plotting-forecast-error-quantiles-against-time-of-day).
+Though existing software solutions (e.g. [NemSight](http://www.nemsight.com.au/), [ez2view](https://www.ez2viewaustralia.info/) and [NEOpoint](https://www.iesys.com/NEO/NEOpoint)) can provide access to some of the same data, most lack a programmatic interface useful for deeper analysis and all are proprietary commercial software. Furthermore, `NEMSEER` adds significant value to users interested in deeper analysis through its documentation. It contains examples showing how users can analyse demand forecast errors and energy price convergence using pre-dispatch demand and price forecast data (obtained using `NEMSEER`) and historical *actual* NEM system and market data (obtained using `NEMOSIS`) [@gormanNEMOSISNEMOpen2018]. \autoref{fig:p5demandforecasterror} is an output of one such [example](https://nemseer.readthedocs.io/en/latest/examples/p5min_demand_forecast_error_2021.html#plotting-forecast-error-quantiles-against-time-of-day).
 
 ![NEM-wide time of day demand error percentiles for 2021 for hour-ahead demand forecasts (i.e. those run during 5-minute pre-dispatch, or 5MPD).\label{fig:p5demandforecasterror}](p5min_error_2021_tod_percentile.png){ width=80% }
 
