### 2023-10-19 #5883 @prakaa
Hi @mfleschutz,

Thanks very much for taking the time to review the package and the paper. I've made changes based on your comments:

---
> So I am starting my proper review.
>
> First thing is noticed when reading through the documentation:
>
> In https://nemseer.readthedocs.io/en/latest/examples/visualising_p5min_demand_forecasts.html there are a lot of annoying warnings:
>
> > WARNING: findfont: Generic family 'sans-serif' not found because none of the following families were found: Source Sans 3, Helvetica, Computer Modern Sans Serif, Avant Garde, sans-serif
>
> @prakaa, can you get rid of them?

**These should no longer be in that example. Sorry about that.**

---
> * [ ]  @prakaa, some of the References in the JOSS paper are no books but have no URL or DOI.

Proof (as in comment above) should now have URLs/DOIs for all references

> * [ ]  In the paper, you do not mention, if there is a similar package out there. This would help to stress the statement of need.
>
>     * [ ]  If there are similar packages, please describe how this software compares to other commonly-used packages(State of the field).

**I have added to the paragraph that follows how `NEMSEER` solves the issues with accessing AEMO forecast data. That paragraph now reads:**

> Though existing software solutions (e.g. NemSight, ez2view and NEOpoint) can provide access
> to some of the same data, most lack a programmatic interface useful for deeper analysis and
> all are proprietary commercial software. Furthermore, NEMSEER adds significant value to users
> interested in deeper analysis through its documentation. It contains examples showing how
> users can analyse demand forecast errors and energy price convergence using pre-dispatch
> demand and price forecast data (obtained using NEMSEER) and historical actual NEM system68
> and market data (obtained using NEMOSIS) (Gorman et al., 2018). Figure 1 is an output of69
> one such example.

---

> I use a MacBook Pro M1 to test. When running the `p5min_demand_forecast_error_2021.ipynb` the `download_raw_data` works. However, calculating the forecast_error (in the cell containing the line `forecast_error = pd.concat(results, axis=0)`) throws the following error multiple times, so I stopped the calculation:
>
> ```
> multiprocessing/process.py", line 108, in run
>     self._target(*self._args, **self._kwargs)
>   File ".../python3.9/multiprocessing/pool.py", line 114, in worker
>     task = get()
>   File ".../python3.9/multiprocessing/queues.py", line 368, in get
>     return _ForkingPickler.loads(res)
> AttributeError: Can't get attribute 'calculate_p5min_demand_forecast_error_simpler' on <module '__main__' (built-in)>
> Traceback (most recent call last):
>   File ".../python3.9/multiprocessing/process.py", line 315, in _bootstrap
>     self.run()
>   File ".../python3.9/multiprocessing/process.py", line 108, in run
>     self._target(*self._args, **self._kwargs)
> ```

**I had originally provided two methods for calculating demand forecast errors. This error is related to the method that uses `xarray`, `pandas` and `multiprocessing`. However, this is a method that many users are unlikely to use to calculate errors across a year (it takes many hours to complete). So I've removed that method, and now only have one method. I have tested this and it works for me. Could you please run the new Jupyter notebook? Note that it write HTML plots to file**
