### 2023-11-06 #5883 @amandadsmith
Regarding **Reproducibility:** The review checklist asks that "If the paper contains original results, results are entirely reproducible by reviewers." It looks like NEMOSIS would be required to reproduce Fig. 1 in the paper. I'm not interested in downloading another package and reproducing it, but not sure whether it's appropriate to check this one off.
