### 2023-11-07 #5883 @amandadsmith
Regarding **Community guidelines:** I'm not sure if there is any guidance provided for "third parties wishing to ... 3) Seek support"
