### 2023-11-07 #5883 @amandadsmith
Regarding **References:** I think the names got mixed up in this one: Hoyer, S., & Joseph, H. (2017). Xarray: N-D labeled arrays and datasets in python. Journal of Open Research Software, 5(1). https://doi.org/10.5334/jors.148
