### 2023-11-07 #5883 @amandadsmith
Regarding documentation: The [quick start guide](https://nemseer.readthedocs.io/en/latest/quick_start.html) says that "AEMO’s [MMS Data Model reports](https://nemweb.com.au/Reports/Current/MMSDataModelReport/Electricity/MMS%20Data%20Model%20Report.htm) describe tables and columns that are available via nemseer." However, the link doesn't directly show information related to the tables--each package in the left sidebar seems to have a list of tables, but it's a lot to browse. If I wanted to know more about e.g. 'CASESOLUTION' would I need to know the package that contains it?
