### 2023-11-07 #5883 @amandadsmith
Also might be nice to explain [here](https://nemseer.readthedocs.io/en/latest/quick_start.html#downloading-raw-data) that the time period is considered as month ending, and will download a file for the entire month corresponding to a given time stamp.

e.g. I queried:
`nemseer.download_raw_data(forecast_type="STPASA", tables="CASESOLUTION", raw_cache = raw_cache, forecasted_start="2010/01/01 00:00", forecasted_end="2010/01/31 11:30", keep_csv=True)`

...and I expected to get one parquet file, one csv file, but got two of each and it just took me a minute to realize why.
