### 2023-11-08 #5883 @mfleschutz
I also finished my review. NEMSEER is a useful and well-documented piece of software with a clear contribution.
@prakaa thanks for your work. @arfon thanks for overseeing this review process.
