### 2023-11-14 2276835
update docs & README as per JOSS review
