### 2023-11-14 deb7fb6
fix xarray ref
