+++ joss-paper/paper.bib
@@ -53,7 +53,7 @@
 
 @article{Hoyer_xarray_N-D_labeled_2017,
   title   = {Xarray: {{N-D}} Labeled Arrays and Datasets in Python},
-  author  = {Hoyer, Stephan and Joseph, Hamman},
+  author  = {Hoyer, Stephan and Hamman, Joseph},
   year    = {2017},
   month   = apr,
   journal = {Journal of Open Research Software},
