### 2023-11-14 #5883 @prakaa
Addressing @amandadsmith's comments here (thanks very much for your time on this review and for your comments!). If this is fine with @amandadsmith, I have incorporated all of the reviewers' feedback @arfon.

---
> Regarding **Reproducibility:** The review checklist asks that "If the paper contains original results, results are entirely reproducible by reviewers." It looks like NEMOSIS would be required to reproduce Fig. 1 in the paper. I'm not interested in downloading another package and reproducing it, but not sure whether it's appropriate to check this one off.
>
> EDITED to add: It seems that the [examples in readthedocs](https://nemseer.readthedocs.io/en/latest/examples.html) also rely on NEMOSIS. If it's expected that the two are always used together, I would mention this in the installation instructions or put a link to NEMOSIS readme on the NEMSEER readme. (If not, can you provide examples demonstrating NEMSEER's functionality alone?)

**I have added a note in the [Installation section](https://github.com/UNSW-CEEM/NEMSEER#installation) of the README that explains how to install NEMOSIS**

---

> Regarding **Community guidelines:** I'm not sure if there is any guidance provided for "third parties wishing to ... 3) Seek support"

**There is now a [Support section](https://github.com/UNSW-CEEM/NEMSEER#support) in the README**

---

> Regarding **References:** I think the names got mixed up in this one: Hoyer, S., & Joseph, H. (2017). Xarray: N-D labeled arrays and datasets in python. Journal of Open Research Software, 5(1). https://doi.org/10.5334/jors.148

**The names should now be correct in the proof above. Thanks for spotting this**

---

> Regarding documentation: The [quick start guide](https://nemseer.readthedocs.io/en/latest/quick_start.html) says that "AEMO’s [MMS Data Model reports](https://nemweb.com.au/Reports/Current/MMSDataModelReport/Electricity/MMS%20Data%20Model%20Report.htm) describe tables and columns that are available via nemseer." However, the link doesn't directly show information related to the tables--each package in the left sidebar seems to have a list of tables, but it's a lot to browse. If I wanted to know more about e.g. 'CASESOLUTION' would I need to know the package that contains it?

**Yes you would and I've provided some more guidance on how to locate the table of interest in a [new subsection](https://nemseer.readthedocs.io/en/latest/quick_start.html#descriptions-of-tables-and-columns) of the quick start guide**

---

> Also might be nice to explain [here](https://nemseer.readthedocs.io/en/latest/quick_start.html#downloading-raw-data) that the time period is considered as month ending, and will download a file for the entire month corresponding to a given time stamp.
>
> e.g. I queried: `nemseer.download_raw_data(forecast_type="STPASA", tables="CASESOLUTION", raw_cache = raw_cache, forecasted_start="2010/01/01 00:00", forecasted_end="2010/01/31 11:30", keep_csv=True)`
>
> ...and I expected to get one parquet file, one csv file, but got two of each and it just took me a minute to realize why.

**I'm not quite sure if this is what you meant, but this is because the function call above calculates a `run_start` and a `run_end` that are in different months. I sort of explained this in the glossary, but I agree with you that it's not clear from anything in the quick start guide. I've added a [footnote](https://nemseer.readthedocs.io/en/latest/quick_start.html#id3) to two places in the quick start guide (one is the section you reference) that explains this behaviour**

---
