+++ .github/workflows/cicd.yml
@@ -119,6 +119,7 @@ jobs:
       # Code coverage to codecov.io
       - uses: codecov/codecov-action@v3
         with:
+          token: ${{ secrets.CODECOV_TOKEN }}
           files: tests/coverage.xml
           flags: unittests # optional
           name: codecov-umbrella # optional
