+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "nemseer"
-version = "1.0.6"
+version = "1.0.7"
 description = "A package for downloading and handling forecasts for the National Electricity Market (NEM) from the Australian Energy Market Operator (AEMO)."
 authors = ["Abhijith Prakash"]
 license = "GPL-3.0-or-later"
