### 2023-11-21 #5883 @prakaa
> @prakaa – looks like we're very close to being done here. I will circle back here next week, but in the meantime, please give your own paper a final read to check for any potential typos etc.
>
> After that, could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
>
>     * The title of the archive is the same as the JOSS paper title
>
>     * That the authors of the archive are the same as the JOSS paper authors
>
>     * I can then move forward with accepting the submission.

Thanks for co-ordinating the review @arfon. DOI is [10.5281/zenodo.10162614](https://doi.org/10.5281/zenodo.10162614)
