### 2023-12-09 #5883 @arfon
> @arfon is the DOI above OK? Please let me know if I need to do anything else.

Yeah, all good. Thanks for the ping. I'm afraid I got a little behind on my JOSS editorial duties 😅
