### 2024-03-22 #5883 @xuanxu
Hey @arfon, the values for version and archive are incorrect, they should be set to `v1.07` and `10.5281/zenodo.10162614` and then reaccept the paper. Can you take care of this?

As [reported here](https://github.com/openjournals/joss/issues/1324) by @sdruskat
